requirements.txt
Flask==2.3.0
pymongo==4.3.3
newspaper3k==0.2.8
beautifulsoup4==4.11.1
requests==2.28.2



